refactor: Lang2SQL 페이지 코드 모듈화
